Move dss.storage.index to dss.index.es.manager
