+++ dss/storage/index_document.py
@@ -20,7 +20,7 @@ from dss import Config, DeploymentStage, ESIndexType, ESDocType
 from dss import Replica
 from dss.storage.hcablobstore import BundleMetadata, BundleFileMetadata
 from dss.storage.identifiers import ObjectIdentifier, BundleFQID, TombstoneID
-from dss.storage.index import IndexManager
+from dss.index.es.manager import IndexManager
 from dss.storage.validator import scrub_index_data
 from dss.util import create_blob_key
 from dss.index.es import ElasticsearchClient, elasticsearch_retry
